Move SrsRtmpToRtcBridge to stream bridge file.
